+++ src/rogallo/document.py
@@ -4,6 +4,7 @@
 # Python imports.
 from dataclasses import dataclass
 from functools import cached_property
+from pathlib import Path
 
 ##############################################################################
 # Gophermap imports.
@@ -12,6 +13,10 @@ from gophermap import GopherMap
 ##############################################################################
 # Port70 imports.
 from port70 import GopherURI
+
+##############################################################################
+# Port79 imports.
+from port79 import FingerURI
 
 ##############################################################################
 # Wasat imports.
@@ -97,6 +102,21 @@ class Document:
     def is_renderable_as_gemtext(self) -> bool:
         """`True` if the document is a source code document, `False` otherwise."""
         return self.is_gemtext or self.is_gophermap or self.is_gopher_error
+
+    @cached_property
+    def suggested_filename(self) -> Path:
+        """The suggested filename for the document, if any."""
+        if (location := self.location) is None:
+            return Path("index.txt")
+        if isinstance(location, Path):
+            return Path(location.name or "index.txt")
+        if isinstance(location, FingerURI):
+            return Path(location.username or "finger").with_suffix(".txt")
+        if location.path and not location.path.endswith("/"):
+            return (candidate := Path(Path(location.path).name)).with_suffix(
+                candidate.suffix or ".txt"
+            )
+        return Path("gophermap.txt" if isinstance(location, GopherURI) else "index.gmi")
 
 
 ### document.py ends here