+++ ChangeLog.md
@@ -6,6 +6,8 @@
 
 - Added a development-oriented tool for taking ANSI "screenshots" of
   Rogallo. ([#336](https://github.com/davep/rogallo/pull/336))
+- Added a `SourceSource` command to save the source of the current document.
+  ([#338](https://github.com/davep/rogallo/pull/338))
 
 ## v1.8.0